backend/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic-settings==2.1.0
httpx==0.26.0
supabase==2.3.4
Pillow==10.2.0
pillow-heif==0.15.0
apscheduler==3.10.4
python-multipart==0.0.9
